refactor: move all job status related functions to shared utils
